+++ portal-service/src/com/liferay/portal/kernel/xuggler/XugglerInstallException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal;
+package com.liferay.portal.kernel.xuggler;
 
 import com.liferay.portal.kernel.exception.PortalException;
